+++ crates/unbase/Cargo.toml
@@ -26,7 +26,6 @@ futures = "0.3.1"
 
 log = "0.4.6"
 sha2 = "0.8.0"
-itertools = "0.8.0"
 serde = "0.9.11"
 serde_derive = "0.9.11"
 serde_json = "0.9.9"
