+++ crates/unbase_test_util/Cargo.toml
@@ -7,6 +7,8 @@ edition = "2018"
 [dependencies]
 log = "0.4.6"
 tracing-subscriber = "0.1.6"
+tracing = "0.1.10"
+itertools = "0.8.0"
 
 [target.'cfg(not(any(target_arch = "wasm32")))'.dependencies]
 env_logger = "0.6.1"
