+++ crates/unbase_test_util/src/lib.rs
@@ -1,5 +1,6 @@
 mod logging;
 mod test;
+pub mod simulator;
 
 pub use logging::init_test_logger;
 pub use test::async_test;
